refactor: refactored mfa component file layout
