+++ web/src/components/Global/MFAHelpers/MfaDialogWrapper.vue
@@ -1,0 +1,14 @@
+<!-- Copyright (c) 2024 t2vee. All rights reserved.
+Use of this source code is governed by an MPL license. -->
+
+<script setup>
+
+</script>
+
+<template>
+
+</template>
+
+<style scoped>
+
+</style>
