+++ web/src/components/Pages/SignInAndSecurity/EditEmailAddress.vue
@@ -17,7 +17,7 @@ import { Label } from '@/components/ui/label/index.js'
 import { Loader, MailCheck, MailX, ArrowBigRightDash } from 'lucide-vue-next'
 import { Button } from '@/components/ui/button/index.js'
 import { DialogClose, DialogFooter } from '@/components/ui/dialog/index.js'
-import MfaCodeInput from '@/components/Global/MfaCodeInput.vue'
+import MfaCodeInput from '@/components/Global/MFAHelpers/MfaCodeInput.vue'
 import axios from 'redaxios'
 import { toast } from 'vue-sonner'
 import { eventBus } from '@/lib/eventBus.js'
