+++ web/src/handlers/PersonalInfo.vue
@@ -4,7 +4,7 @@ import {CardDescription, CardTitle} from '@/components/ui/card/index.js'
 import {BookType, CalendarFold, CircleUserRound, DoorOpen, Earth, UserRound} from 'lucide-vue-next'
 import EditPublicProfile from "@/components/Pages/PersonalInfo/EditPublicProfile.vue"
 import EditBirthdayDialog from '@/components/Pages/PersonalInfo/EditBirthdayDialog.vue'
-import EditDetailDialog from '@/components/Global/EditDetailDialog.vue'
+import EditDetailDialog from '@/components/Global/MFAHelpers/EditDetailDialog.vue'
 import EditUsername from '@/components/Pages/PersonalInfo/EditUsername.vue'
 import EditRegionalSettings from '@/components/Pages/PersonalInfo/EditRegionalSettings.vue'
 import EditLocaleSettings from '@/components/Pages/PersonalInfo/EditLocaleSettings.vue'
@@ -64,9 +64,7 @@ function expandLocale(shortLocale) {
     <div class="flex gap-4 mt-4">
       <EditDetailDialog
         title="Country/Region"
-        :desc="
-          userData.profile?.address?.country ?? userData.profile?.address?.country ?? 'Not Set'
-        "
+        :desc="userData.profile?.address?.country ?? userData.profile?.address?.country ?? 'Not Set'"
         :icon="Earth"
         :dialog-page="EditRegionalSettings"
       />
