+++ web/src/handlers/SignIn&Security.vue
@@ -5,7 +5,7 @@ import {CircleEllipsis, Fingerprint, KeyRound, MailCheck, MailX, Phone, PhoneMis
 
 import AddPhoneNumberDialog from '@/components/Pages/SignInAndSecurity/AddPhoneNumberDialog.vue'
 import EditPhoneNumberDialog from '@/components/Pages/SignInAndSecurity/RemovePhoneNumberDialog.vue'
-import EditDetailDialog from '@/components/Global/EditDetailDialog.vue'
+import EditDetailDialog from '@/components/Global/MFAHelpers/EditDetailDialog.vue'
 import EditEmailAddress from '@/components/Pages/SignInAndSecurity/EditEmailAddress.vue'
 import EditMfaMethods from '@/components/Pages/SignInAndSecurity/EditMfaMethods.vue'
 import UpdatePasswordDialog from '@/components/Pages/SignInAndSecurity/UpdatePasswordDialog.vue'
